SVG overlay: respect broken slices when rendering edges
When an edge to a (potentially virtual) child and parent in the next
or previous section is rendered, broken slices are now taken into
account. So far the edge starting from the current (potentially virtual)
node, was only drawn to the intersection in the adjacent section,
regardless of whether it was broken or not. This caused a visual brake
betwen the edge segments in the current and the adjacent section. Now,
edges are drawn so that they don't suggest a node in a broken section,
if no real node (unless there actually is a real node).