Stack: floor() instead of round() in project to stack Z conversion
Stack space is discrete in Z and so far CATMAID rounded coordinates in Z
when converting from project to stack space. To be consistent with the
neuron tracing part of CATMAID, Z is now set to lower integer instead to
the closest one. This means a coordinate between two sections is always
considered part of the first one.

If not changed, this inconsistency is a problem for orthogonal views:
Spatial annotations (e.g. treenodes, connectors) are queried and
displayed for a section if they are on it or between this section and
the next one. Project coordinates have to be converted to stack space to
create the query bounding box and to tell how far away a node is from
the current section for a visibility test. With rounding, there were
situations were some nodes would not be displayed at all: if the node is
between two sections, close to the next, it was so far mapped to the
next section. It was therefore hidden for the current section, but also
not queries for the next one (since it comes before it in Z, which also
would have prevented its display).

The alternative would have been to keep the rounding, but to adjust the
query bounding box to look half a section back and half a section ahead,
and like wise change the visibility tests for nodes.