3D viewer: fix z-plane for ortho stacks and viewer
Z-planes in the 3D viewer now show the plane of the focused stack
viewer. If the focused stack viewer is not an XY stack, it shows the
appropriate plane for the viewer.

Make WebGLApplication listen to stack viewer focus change events and
adjust the static content as necessary. All static content is adjusted
rather than just the z-plane because other static content will need to
adapt to ortho stacks and stack viewers also.

See catmaid/CATMAID#1048.