SVG overlay: return independent promise for location of real node
So far the submitter was returned, which can act as a promise. However,
it acts as a shared promise which could change state when updated from
other code parts. To have an independent promise for the location, which
nevertheless respects the submitter, the call to the submitter is
wrapped in its own promise.