Change license from AGPLv3 to GPLv3
This move was discussed with all previous contributors and agreed on.

Third parties showed interest in contributing to CATMAID if we move away
from AGPLv3 to a license that doesn't protect CATMAID's back-end. This
resulted from their fear of being forced to expose back-end
infrastructure. As a client server application, CATMAID's server (back-end)
is never actually distributed to users in a running CATMAID instance.
The GPLv3 doesn't enforce free and open access to changes done to the
back-end as long as the server isn't distributed to a user. AGPLv3 was
chosen before to explicitly enforce the publication of changes to the
server. In effect, closed-source changes can now be done to the back-end
as long as the server code isn't distributed to others.