Stack viewer: add optional navigation offset
Add an optional navigation offset to stack viewer such that whenever the
project moves to a location, the stack viewer (if it is navigating with
the project) moves to that location plus the offset translation. Stack
tiles, tracing overlays, etc. all work correctly at the stack viewer's
location.

This is especially useful in segmentation for viewer adjacent sections,
and was originally implemented in
catsop/CATMAID@308543b2261648feb8822cde282f98551518ba3d. It may be
useful in mainline CATMAID, but more importantly having it here makes
sure CATMAID respects this offset by accessing coordinates and
initiating navigation in the intended ways.