Tracing tool: add measurement action
Add an action, bound to the X key, that starts measuring distance from
the current cursor position in the stack. The radius tool is show while
distance is measured in all three axes in both stack space and project
space. On click, the tool is removed and the final distance from the
starting point to the click is shown in the status bar.

Closes catmaid/CATMAID#798.