Improve performance of conenctor node queries
Connectors are queried for all treenodes found in a given bounding box.
These lists of treenodes can get quite big and so far the query
condition was expressed as

  SELECT * FROM treenode_connector tc
  WHERE tc.treenode_id IN (id1, id2, id3, ...);

As it turns out, Postgres can't optimize this very well. IN is basically
an equicalend to OR, each comparison is evaluated until the result is
ready. If rewritten as

  SELECT * FROM treenode_connector tc
  INNER JOIN (VALUES (id1), (id2), (id3), ...) vals(v) ON tc.treenode_id = v;

Postgres can optimize the query. With the query used so far, the
conenctors for 4000 were retrieved in 120ms. With the new query, only
10ms are needed. There were about 15 million entries in the treenode
table.