Neuron search: register expanded neurons with name service
When an annotation is expanded, all the neurons are listed that are
annotated with it. Their neuron names could be displayed as "null" if
the name service hasn't seen this neuron before. To make sure the
correct information is available, the query used when expanding an
annotation will now also register the resulting neurons with the name
service.

Fixes catmaid/CATMAID#1057