Split/merge dialog: destroy complete 3D viewer on close, not only space
This fixes a bug where an already closed split/merge dialog was still
registered to a stack view change event. Handling this event would throw
an error, because the 3D viewer's space was already destroyed. Now that
the 3D viewer is destroyed properly, it is also unregistered from this
event.