Review widget: fix virtual step calculation for nodes close to top
So far the calculation for steps between one node and the next in the
review widget used the vector between the current and the next node.
This vector was transformed into stack space and it was checked if a
section had to be skipped. Unfortunately, this approach is flawed,
because the project-to-stack transform respects the stack translation.
To fix this, both the current and next position are transformed to stack
space individually and then the difference is used.