3D viewer: listen to location changes and update Z plane accordingly
Regardless of how the location is changed in the project, the 3D viewer
will now be notified. For now this will only cause the z plane to
update. This change allowed also the removal of a tightly coupled update
call on the 3D viewer in the tracing tool and its counterpart in 3D
viewer which updated *all* known 3D viewer instances at once.

See catmaid/CATMAID#1061