Improve performance of PostGIS treenode query further
So far we only used te &&& operator to test an edge against the query bounding
box (the line string). This is relatively fast, but returns quite some false
positives, because only the bounding box of edges is tested for intersection,
not the edge itself. To reduce the result set ST_3DDWithin can be used. It
allows to limit the result set by a distance to another geometry. For us it can
be used to only allows edges that are no farther away than half the height of
the query bounding box from a plane that cuts the query bounding box in half in
Z. While returning much fewer false positives, it is slow compared to &&&.
However, on the PostGIS mailing list it was suggested that bot &&& and
ST_3DDWithin can be used together. This combines the best of both approaches by
having the speed of &&& and much fewer false positives of ST_3DDWithin. Still,
only a n-d index is used (according to the query plan ST_3DDWiting wouldn't pick
up a 2-d index in this query even if present.

In denser regions this reduced the result set size by 50%.