Improve performance of missing treenode query
This query gets treenodes that are not in the query region, but are
referenced by connectors. Usually, the list of such nodes is small. But
to stay consistent and be prepared for larger list, the IN operator is
replaced with a join on a VALUES list, just like for the other queries
involved in retrieving treenodes.