Split/merge dialog: add input/output visibility and shading controls
This adds three more controls to the split/merge dialog. They control
weather inputs and outputs are displayed in the embedded 3D viewer and
which shading method is used. By default, "active node split" is used.
This can be changed to show "Strahler index" coloring instead, which
helps with depth perception.

Closes catmaid/CATMAID#963