Prevent accidental page reload and history browsing
If attempted, a warning will be shown. It makes clear that window layout and
content in CATMAID won't be saved if the user continues. The actual display and
options given to the user is up to the browser. Chrome will ask users if they
want to stay or reload/leave.

Fixes catmaid/CATMAID#1066