Tracing tool: clear neuron name in top bar if neuron is deselected
(De)selecting a neuron triggers an active node changed event. In case of
no new active node, all fields of the active treenode are set null.
This commit makes sure the tracing tool will not only clear the top bar
if a falsy new active treenode is passed to the handler, but also if the
object's ID property is null.