Add an optional profiling middleware
This middleware can be used to get a cProfile profile for requests on
individual views. It is not enabled by default, but it can be easily
switched on for development purposes. To do so, just append it to the
list of used middlewares in a settings.py file:

  MIDDLEWARE_CLASSES += ('catmaid.middleware.ProfilingMiddleware', )

Then, for each request that contains a field named 'profile', a
cProfile log is recorded and stored to a file in /tmp. The file name
follows the pattern 'catmaid-$(host-address)-$(timestamp).profile'.

This is useful to find out which part of a request takes what time.