SVG overlay: make node event identity check more robust
According to our server side client error log it happens from time to
time that a mouse up event is triggered without the node being actually
part anymore of the SVG element. This in turn causes the node identity
check to fail and raise an error. To handle this situation more
gracefully, it is now checked for this situation and no error is
displayed to the user.

There is a race condition when node are updated and events are
propagated. I was able to reproduce this sporadically by selecting a
node and then scrolling a couple of sections before releasing the mouse
button.