Fix treenode update trigger for updating edges
An edge is updated if either the location of a node or its parent
changed. For testing the latter I used OLD.parent_id != NEW.parent_id.
This is fine as long neither OLD.parent_id is NULL nor NEW.parent_id is
NULL, because the != operator doesn't work with NULL values (it yields
null). To fix this, OLD.parent_id IS DISTINCT FROM NEW.parent_id is now
used.

Fixes catmaid/CATMAID#1070