Improve query performance for listing all annotations
If all annotations are requested, without any constraints, a simple SQL
query is now performed instead of utilizing Django's ORM overkill. This
improves performance on my test system from about 350 ms to 30 ms.

If the annotation result set is constrained, Django's ORM is still used.
But this is likely to change in future updates, too.