Fix performance regression in edge update trigger
Fix a performance regression in the trigger function for updating
treenodes. IS NOT DICTINCT FROM on a primary key results in a sequential
scan rather than an index. Since only the queried value can be null, not
the key itself, simplify this to a NULL check and an index scan.

Diagnosed with @tomka.

Closes catmaid/CATMAID#1072.