Change review status from percentage to counts
Change the review status API to return a tuple of the total number of
nodes and number of reviewed nodes for each skeleton rather than just a
review percentage. This allows the API to be reused for more generic
purposes, such as front-ends that compute weighted aggregate review
statistics.

Update widgets as appropriate.