Connectivity widget: load reviews asynchronously
Previously the connectivity widget backend queried and returned review
status for all partners' reviews, team summary, and union summary. When
selecting a review filter in the widget, the entire widget was redrawn.

Instead of redrawing the entire widget (which is by far the slowest
operation), only update the review column when selecting a review
filter. Since this column can be updated on demand, there is no reason
to send all possible review summaries from the backend. Instead the
connectivity backend returns only a list of reviewers with any reviews
for the partner set, and the connectivity widget reuses the review
status API.

Because review status is not queried until the table is created and
DataTables must then re-process the table data, this erodes some of the
time-to-first-draw improvements in 812c7c1. However, changing reviewers
is much faster, and connectivity API call itself is ~%40 faster.