SVG overlay: speed up navigation to child
Speed up navigating to the child of the active node by creating a
specific backend API rather than reusing the next branch API. This API
returns the same format as the next branch API, only without array
entries for next node of interest and next branch point, so that it can
reuse the branch cache in the SVG overlay.

Branch caching logic and branch node cycling now check that cached
branch information contains the specific node type required, which
should make this extensible if future next-node-of-interest-in-topology
types are added.