3D viewer: use specified synapse color for restricted connectors
Eventually, the restricted connector's extra geometry should be removed,
so that we don't need extra coloring anymore. We would then just need to
show only the visible restricted connectors and hide the rest. This will
be done in a separate commit to have the underlying issue closed with a
rather small change first.

Closes catmaid/CATMAID#753