SVG overlay: extend virtual node ID to include location
To be able to deduce the location of a virtual node in a stack viewer
other than the one in which the virtual node was created, this extra
information has to be stored somewhere. Eventually, this will be a
shared data structure, but for for now this is taken care of by adding
the complete project space location to the virtual node's ID.

This also fixes a bug where selecting a virtual node in one view,
selecting another stack view and navigating with brackets threw an
error.