Navigator tool: fix de/register() if button is not available
A button ID passed through the Navigator tool's register and unregister
method might not be valid or the button is not created on purpose (i.e.
if the user doesn't have permission to see a tool). Therefore, do check
  if the DOM element actually exists before setting properties of it.

Before this caused an error if one opens a tracing view link as a user
who can't see the tracing tool.