Init: periodically check that client and server are same version
The client now checks every 15 minutes whether it is the same CATMAID
version (via git describe) as the server. If not, an error dialog is
shown prompting the user to refresh their browser.

This reduces the chance that a user can unknowningly use the wrong
client version, which can lead to buggy behavior and potential data
loss.