Review widget: fix step limiting
So far the step limiting checked if the new target section has a higher
index than the original target. If this was the case, it claimed the
step doesn't need to be limited. This is clearly wrong and only works if
moving from lower Z indices to higher ones. I suspect that this caused
the problem @ceesem saw: only W would move to virtual nodes. Limiting is
now done by incrementing sections to the first section that is a) no
broken slice and b) as close as possible to the number of allowed steps.
If one reaches the original target node while looking for the next valid
section, it is used and the limiting is canceled.

Closes catmaid/CATMAID#1089