Skeleton projection layer: support Strahler shading
This Strahler shading pretends the current node to be the root and uses
basically separate Strahler numbers for upstream and downstream nodes.
This is convenient to only show part of the skeleton. Two modes are available:
show projection until a maximum Strahler number is reached, either with
transparency or without.