Skeleton projection layer: support for both min and max Strahler constraints
For the currently implemented relative Strahler display this limits the
display to nodes that have the same Strahler number with respect to the
current node. In gradient mode, the gradient spans the range from min to
max with max being limited to what is actually available on the
skeleton. Both can be disabled when -1 is set to them.