Make JsonResultHandler call error handler also in silent mode
A handler created with createJsonResponseHandler can be passed a silent
flag which will instruct it to not show any dialog in case of an error.
If this flag was set, so far error handlers wouldn't be called due to
short cut logic. Swapping error handler call and flag test makes sure
the error handler is run all the time.