Make json handler silent for node deletion operations
Node deletion returns a promise and similar to exceptions, errors can be
dealt with by the caller. It uses the json result handler, but there is
no need to have it show an extra error message in case of errors,
because the error can be dealt with separately. This also fixes a
problem where an error on node deletion was displayed twice.

Closes catmaid/CATMAID#1067