Neuron name service: fix refresh
Since input skeletons are respected by a name service update, refresh()
wasn't working anymore. It provided null as the list of skeleton ids to
update, meaning to update all. Instead of only checking only for
undefined as list of skeleton IDs, everything falsy as a list of
skeleton IDs will cause the update of all names.