Support easier skeleton merging with preservation of all annotations
Instead of requiring a caller to name all annotations that should be
kept, the back-end API to merge neurons also accepts a request without
an annotation_set parameter. If it is omitted, the winning skeleton will
have all annotations it had before plus all annotations from the losing
skeleton.

This is useful for fixing issues like catmaid/CATMAID#1102.