SVG overlay: merge all annotations of single-node skeletons, if possible
If not disabled, single-node skeletons are merged without asking if it
is joined in another skeleton. All its annotations are moved over. To
not require loading all annotations for both skeletons in such a
situation, a new feature of the skeleton merge API is used: leaving out
the annotation_set parameter will cause all annotations of both winning
and losing skeleton to be linked to the winning skeleton after the
merge.

See catmaid/CATMAID#1102