SVG overlay: make close node detection ignore nodes in section ahead
There has recently been a change which made project to stack coordinate
conversion use floor() instead of round().  This maps all nodes between
two section to the one with the lower Z index. Therefore, excluding
nodes that are not on a given section has to be done by testing if
relative stack space distance between 0 and 1.