Make inclusion of non-pipeline files easier
There is now the non_pipeline_js dictionary in the pipelinefiles module
which holds an identifier and a file that should not be included through
pipeline. If files are added there, they will be made available through
the provided path under STATIC_URL. No other changes are required to add
such a file.