SVG overlay: prevent connector info retrieval for deleted edges
When a link between a connector and a treenode is removed, so far an
error was generated when one moves with the mouse over the (deleted)
edge before it a screen update was done. This happened due to
information for this link was tried to retrieve. This failed, because
the link was removed already. To prevent this, the edge representing to
link gets a 'suspended' flag when it is removed, which is reset when the
request returns. Edge info will not be retrieved if this flag is set.

Closes catmaid/CATMAID#1101