Connectivity matrix: allow descending sorting for all sorting modes
All row and column sorting defaults now to ascending. With a separate
option for both rows and columns, a selected sorting mode can now be
made descending. The number of elements in the sorting mode drop-down
menus is reduced by this. This is done in preparation to add new
sorting modes like discussed in catmaid/CATMAID#1116.