Selection table: let data table create rows
Instead of creating the HTML table rows manually and let DataTable read
it, a data source with all needed information is now passed directly to
the Datatable initialization. Besides being faster, this also allows for
improvements like not being required to redraw the complete table when
new skeletons are added.