Selection table: propagate DataTables sorting back to widget
The selection table manages its skeletons with an internal list of
skeleton models and an object mapping skeleton IDs to these models.
Before DataTable was used, this list was always following sorted
according to the user's selection. This sorting was expected to be in
sync with the displayed table when skeletons where colored through the
"Colorize" button. This commit makes sure the sorting selected from the
displayed DataTable is propagated back to the widget's internal list.
This makes sure the list is consistent with the displayed table and
coloring works like before.

Fixed catmaid/CATMAID#1143.