Selection table: update all rows on model changes
DataTables uses a cache to not recreate row content every time it is
needed (e.g. on page change). This cache has to be invalidated when the
underlying model changes. Some of the invalidation calls could be made
more efficient by only invalidating affected rows. This will fix the UI
problem referred to in catmaid/CATMAID#1146