Graph widget: fix subgraph creation
When subgraphs (like axon/dendrite split) are created, new nodes will be
added to Cytoscape's canvas. These nodes will be styled with the same
CSS that is used for all other nodes and therefore some data is expected
to be available on a node. In this case, our CSS expects nodes to have
a shape property. Previous versions of didn't fail when it wasn't
available, but new versions will raise an error. Adding the expected
shape property to new nodes created for subgraphs, fixes this problem
and nodes can be split again.

Fixes catmaid/CATMAID#1142.