3D viewer: improve SVG export progress info
The changes done to fix catmaid/CATMAID#1133 work in many cases.
However, the browser still did not have many chances to queue rendering
of progress information at the start. The changes in this commit will give the
browser more opportunities to queue rendering of progress information in
its event loop right from the beginning of the export.