3D viewer: shade by planar active node distance
Add a shading option to shade skeletons by the distance from some plane
through the active node. This plane can either be parallel with the
z-plane of the project space or with a camera view plane (i.e., a plane
whose normal is the ray from the camera to the active node).

To support this, create a custom material wrapping Three.js'
LineBasicMaterial with defined GLSL insertion points to allow easy
shader manipulation, like modifying fragment color, without needing to
recreate an entire shader pipeline.

Closes catmaid/CATMAID#893.