Selection table: don't copy filtered list from table to widget
To reflect the table sorting in the widget's internal skeleton list
representation, the DataTable writes back its list to the widget when
its ordering changes. This should of course not happen if the DataTable
contains a different dataset than the widget, which can happen with the
current implementation. Therefore, this commit adds some checks that
makes sure a skeleton list is only written back if the data in the
DataTable is the same as in the widget.

Fixes catmaid/CATMAID#1160