Selection table: populate DataTable always with complete list
So far the DataTable display used the filtered list of skeletons as
source. Since DataTable instances can filter on their own, the full list
of skeletons is now used. This makes initialization cleaner and allows
for further optimizations to now recreate the DataTable with every
update.

This enables also a non-regex filter, if no "/" is prepended to the
filter term.