Fix ID generation of InstanceRegistry
CATMAID's InstanceRegistry is used to manage instances of individual
widgets. It assigns the lowest possible integer as widgetID to each new
widget instance. Finding the lowest valid ID did not work in all cases,
which would usually lead to chaos in new widgets. Such a situation could
be created by opening three widgets of a kind (e.g. Selction Table) and
closing the first instance. A new widget of this kind would not get ID
1, as it would be expected, but 2, which is used already. The test if an
ID was used already did not check if is a field in the internal
instances object. Instead it checked if a potential new ID was an
available index in an array of existing IDs. In the example above the
lowest available index was 2 (the array of existind IDs after removing
the first was [2, 3]), so it was used as new ID, which of course isn't
the right thing. This is now fixed and a unit test is added to make sure
this bug doesn't show up again.

Fixes catmaid/CATMAID#1166