Add DVID middleware and models
This middleware and the mapping models allows CATMAID to be used as a
front-end for tiled image data stored in DVID. It can be activated and
configured by adding the following to setting.py:

  MIDDLEWARE_CLASSES += ('catmaid.middleware.DVIDMiddleware',)
  DVID_URL = 'http://my.dvid.server:7000/api'
  DVID_FORMAT = 'jpg:80'
  DVID_SHOW_NONDISPLAYABLE_REPOS = False

Once active, CATMAID's project menu will be populated with available
repositories on the DVID server. For now, only data instances of the
type 'imagetile' will be treated as a displayable stack. The setting
DVID_SHOW_NONDISPLAYABLE_REPOS controls if repos should be displayed
even without such a data instance.

See catmaid/CATMAID#1081