Ensure projects have all expected classes and relations
After every database migration run (manage.py migrate) it is checked if
all projects have the classes and relations they expect. If a project is
added the same thing is done for the new instance. This allows default
classes and relations to be available for projects.

There is currently only one default class: stackgroup. There are two
default relations: has_view, has_channel.