Stack groups: add admin tools and proxy models
The proxy models StackGroup and StackStackGroup shadow ClassInstance and
StackClassInstance, respectively. Their purpose is to pre-filter queries
for ClassInstance objects to only those of class 'stackgroup'. With the
new admin inlines, stacks can be added to stack groups from within an
inline form on the stack edit/add page. On the stack group add/edit
page, stacks can be linked as well.