Add CATMAID.fetch() method to wrap a request in a promise
With CATMAID.fetch(url, method, data) a request is queued in CATMAID's
request queue and a promise is returned. If the request was successful
and no error field is set in the returned JSON, the promise is resolved.
Otherwise, it is rejected. The URL passed to this function must be
relative to django_url/CATMAID_URL.