Add basic stack group loading
If a stack group is loaded from the Project menu, all stacks linked to
it will now be loaded one after the other. If a stack is linked as a
view, it is opened in its own stack viewer. If it is linked as a
channel, it will be added to the first stack viewer created for the
group.