Stack groups: auto-layout after loading
After a stack group has been loaded, a layout function is called. This
function currently only changes the window layout if there are three or
four windows of which three are stack viewers, each with a primary stack
of different orientation. If there is no fourth window, a selection
table is created. These four windows then are laid out so that they all
have the same size, that the XY window is on the left side on top of
the XZ window and that the ZY window is on the right on top of the
fourth window.

Closes catmaid/CATMAID#1082