Stack groups: use key shortcut window for multi-view layout
The previously used selection table is part of the tracing tool and is
therefore no good option to be opened alongside a multi-view stack group
during auto-layout. At this point, no tool is active and the selection
table isn't available in principle. Instead, the keyboard shortcut
window is now used.