Do project validation only after CATMAID migrations
Project validation expects of course the project database tables to be
present. Therefore, it is not the correct thing to run this test after
every the South migration of *every* Django application, but only after
the migration of catmaid itself.

Also move both project and ontology dummy project test in one function
that makes sure both are run only after the migration of CATMAID.

Fixes a problem introduced with be76376. Thanks @aschampion for the
pointer.