Tests: fix insertion tests by adding admin account
CATMAID expects now one admin account to be present to be able to create
missing classes and relations for projects. This is already the case for
all tests that use the testdata fixture, but not for the insertion test,
which didn't use any fixture so far. This commit adds a minimal CATMAID
environment fixture by adding a single admin user. This fixture is now
used by the insertion test case.