Selection table: improve UI update performance
To trigger an update for the table rows of specific skeletons, their ID
can now be passed to the invalidate() function. If a list of skeleton
IDs was passed to it, only those rows matching the passed IDs will be
updated. Especially for large lists (e.g. all locked neurons), this is
an improvement if one operates on smaller sets of neurons.