Add workaround for Three.js issue 7361
Current Three.js (r72 and r73) have a bug that doesn't allow us to
change vertex coloring of geometries dynamically. This is discussed in
issue 7361 of their tracker on GitHub. This commit adds a temporary
workaround for this by intercepting the "colors" field of Geometry
instances. The code for this is isolated and be removed dynamically.