3D viewer: fix wrong depth sorting in orthographic mode
The wrong sorting of rendered skeletons and their components is somehow
linked to the logarithmic depth buffer. It is still unclear what the
underlying problem is, but it is unlikely a binning problem. In
Three.js, the only way to change the Z buffer type for a renderer is to
create a new one. This is why this refactor moves the renderer
initialization into its own function. For now, every time the camera
mode is changed, a new renderer is created.

Fixes catmaid/CATMAID#1168