Skeleton projection layer: support functions to generate color
If a shading method is not a function, but an object with the fields
"opacity" and "color", its color function will be used to generate the
color for each node and edge drawn. The opacity method is the same as if
no object was used. Existing methods will therefore continue for now to
only define the opacity method.