Tracing tool: allow other layers to be used for closest neuron search
If "g" is pressed, the tracing tool will now check all layers of the
active stack viewer from top to bottom if they have a method called
"getClosestNode". If they do and it returns a node ID, this node ID is
selected. Otherwise the regular mechanism of SVG overlay are used. This
will allow the skeleton projection layer to be used for node selection.