Events: refactor to add explicit constructor and mixin functions
The behavior didn't change, but the name and the internal organization.
CATMAID.EventSource can now also be used as a constructor to create
stand-alone event sources. Additionally, using CATMAID.EventSource as a
mixin became more readable with CATMAID.asEventSource(object)---which
now all places use that use event source functionality.

Note: I suggest that we stick to this naming scheme for functions that
are intended to be used as mixins.