Remove annotated entity query with DataTables response
There is no real need for a back-end method that returns DataTables
formatted results. DataTables can deal with the response of the regular
API in the front-end. To not maintain two versions of the same
functionality, the DataTables special case is removed. For this to work,
the Neuron Navigator's neuron list display (based on DataTables) had to
be adapted. The required functionality has been merged into the regular
API (sorting, ranges, name filter, total records).

This makes the API cleaner. Documentation and parameter cleanup will
follow.