Selection table: lower memory foot print and improve performance
Instead of constructing a new DataTable instance for every update,
create DataTable instance once and then update its content. This
relatively easy through DataTable's 'ajax' option which essentially
allows for a function to supply the data. The function is now executed
with every update. It gets all filtered skeletons and supplies them
along with extra information to the DataTable instance.