Selection table: add annotation filtering
Annotation filtering is very useful in a Selection Table as a user can
make use of existing groupings. To implement this, the DataTable
instance had to be used in a slight different way than before. All the
sorting, filtering and pagination is now done in the widget and not by
the DataTable instance. This gives us much more flexibility for
filtering.

Closes catmaid/CATMAID#1119