Volumes: add non-intersection warning setting and handler
New nodes can now be tested against intersecting a certain volume. The
Tracing section of the settings widget allows to choose a volume against
which new nodes will be tested. If they are outside of it, a warning
will be shown.

Volumes them self can be added through the Volume Manager, a new widget
available through a new box icon, last in the list of tracing tool
widgets.