Volumes: test bounding box overlap as intersection test
This of course works only for box geometries, but this is all we need at
the moment. ST_3DIntersects didn't count it as intersection if the query
point is inside the volume geometry, something we have to investigate
further. So in order to make this work at least with boxes, the &&&
operator is used.