Selection Table: fix review percentage display after clear
Clearing a Selection Table causes the review field to be set to a new
object, a previous instance wouldn't get updated anymore. Therefore,
keeping references to this field causes problems after clearing the
table. This is what broke the review percentage display. Instead of
keeping an old reference, the actual field of the widget is now used
when populating the table.

Fixes catmaid/CATMAID#1201