Annotations: allow and-combined or-queries
Each annotation field, starting either with 'neuron_query_by_annotation'
or 'neuron_query_include_subannotation', can now not only include a
single annotation ID, but a list of annotation IDs. If this is the case,
all annotations in one field are treated as OR combination while the
combination of individual fields remains to be AND combined. This also
works for sub-annotation inclusion. If sub-annotations should be used
for a set, sub-annotations for each individual set element are looked
up.

This is required to allow the front-end more complex search queries.