Search widget: support regex and partial name for annotation query
So far the 'annotated' field of the search widget required the use of
a single existing annotation. This is very limiting, especially if one
wants to search for entities annotated with annotations that share a
common prefix or suffix.

This commit changes this behavior so that a single existing annotation
is used for the query if it exists (i.e. it has the name entered in the
input field). If this is not the case, the annotation string is treated
as a regular expression on annotation names if it starts with a slash
character ('/'), otherwise it is treated as a regular search pattern
over all annotations. For instance, finding all annotations that are are
annotated by either a1 or b2 would look like '/a1|b2'. This also works
with sub-annotation queries.