Tracing tool: add shortcut to find nodes by label
Add a shortcut (\) to go the nearest node with a label matching a regex.
If a skeleton is active, this is limited to the active skeleton and goes
to the nearest node by path distance. Otherwise it goes to the nearest
node with a matching label in project space (a optional distance limit
is available in this mode). Shift+\ cycles through results. Ctrl+\
repeats the last search.