Volume manager: add basic box visualization
To visualize edited box volumes, the box tool's layer is used. This is
currently more of a hack to get an idea of the extent of a box. In its
current state, the box tool, ROI tool and its layers can't be used with
ortho stacks (see catmaid/CATMAID#1073), which should change soon. They
also don't have a good way to represent Z in the UI. However, it
provides a simple way to show project space based boxes as a layer.