DVID stack wizard: add option to create new (ortho) project
This option allows to create a new project for the newly created stack.
For now it has the same name as the stack. By default the stack is
linked as XY view, but optionally an ortho-view project can be created.
In this case, the stack is linked three times to the project with every
link having a different orientation. This allows for very easy creation
of DVID based ortho-projects.