DVID stack wizard: create three stacks if ortho views requested
CATMAID currently doesn't allow the same stack linked twice to a
project. The database isn't enforcing this, but the front-end raises an
error. To make the DVID stack import wizard play along with this, it has
to create three stacks if ortho-views are wanted, instead of linking one
three times.