Fix segfault on cropping TIFF stacks with pgmagick and GeoDjango
This adds a hack to disable GDAL support in GeoDjango. It seems there is no
easier way due to the GDAL_LIBRARY_PATH setting not behaving as documented
(i.e. setting it to a non-existant file will disable GDAL). Disabling GDAL is
unfortunately needed, because it breaks TIFF support for pgmagick, which is
required by e.g. the cropping back-end. GeoDjango loads GDAL into memory
which a) is a waste of memory if we don't use it and b) on at least Ubuntu
GDAL seems to be compiled in a way that it takes over TIFF I/O. And this
causes pgmagick to segfault. This patch is added here, because the database
engine is the first thing that is loaded which loads GeoDjango's GDAL module.
The patch is based on the GeoDjango's GDAL module for Django 1.6:

https://github.com/django/django/blob/stable/1.6.x/django/contrib/gis/gdal/libgdal.py

It temporarily changes the operating system name so that GeoDjango
raises an error that is caught by GeoDjango and will cause GDAL to be
disabled.

Fixes catmaid/CATMAID#1218