Cropping: fix single- vs. multi-channel selection in front-end
The last parameter of CATMAID.crop() allows to select if the output file
should contain only one channel of if it should be an RGB file. The
latter is useful for e.g. multi-channel output if different stacks
represent different channels. However, the selection was inverted by
accident and should work now correctly.